celery-queue/requirements.txt
amqp==2.2.2
Babel==2.5.3
billiard==3.5.0.3
celery==4.2.1
flower==0.9.2
kombu==4.2.0
pytz==2018.3
redis==2.10.6
tornado==5.0.2
vine==1.1.4
scikit-learn
scipy
numpy
pandas
tensorflow
flask-cors
opencv-python
Pillow
torch
torchvision
scikit-image
flask-socketio
eventlet
watchdog
psutil
inotify